requirements.txt
# Core dependencies
streamlit==1.22.0
torch==2.0.1
torchvision==0.15.2
opencv-python==4.7.0.72
numpy==1.24.3
matplotlib==3.7.1
Pillow==9.5.0 